Update go_env.py
made window height and width integers

+++ gym_go/envs/go_env.py
@@ -176,8 +176,8 @@ class GoEnv(gym.Env):
             from pyglet.window import key
 
             screen = pyglet.window.get_platform().get_default_display().get_default_screen()
-            window_width = min(screen.width, screen.height) * 2 / 3
-            window_height = window_width * 1.2
+            window_width = int(min(screen.width, screen.height) * 2 / 3)
+            window_height = int(window_width * 1.2)
             window = pyglet.window.Window(window_width, window_height)
 
             self.window = window
